SRB2: merge 'fix' for broken identical colormap checking to trunk
git-svn-id: svn://code.srb2.org/SRB2/trunk@5258 d18046d3-e9af-4449-a702-ac2b6c676ba2

+++ src/r_data.c
@@ -680,7 +680,7 @@ INT32 R_CreateColormap(char *p1, char *p2, char *p3)
 			continue;
 		if (maskcolor == extra_colormaps[i].maskcolor
 			&& fadecolor == extra_colormaps[i].fadecolor
-			&& maskamt == extra_colormaps[i].maskamt
+			&& (float)maskamt == (float)extra_colormaps[i].maskamt
 			&& fadestart == extra_colormaps[i].fadestart
 			&& fadeend == extra_colormaps[i].fadeend
 			&& fog == extra_colormaps[i].fog)
@@ -827,7 +827,7 @@ void R_CreateColormap2(char *p1, char *p2, char *p3)
 			continue;
 		if (maskcolor == extra_colormaps[i].maskcolor
 			&& fadecolor == extra_colormaps[i].fadecolor
-			&& maskamt == extra_colormaps[i].maskamt
+			&& (float)maskamt == (float)extra_colormaps[i].maskamt
 			&& fadestart == extra_colormaps[i].fadestart
 			&& fadeend == extra_colormaps[i].fadeend
 			&& fog == extra_colormaps[i].fog)
